Fix signed overflow in proc allocation and unsafe strcat in dsock
- Change static int size to size_t in alloc_lproc() to prevent signed
  integer overflow when growing the process table
- Replace unbounded strcat() with bounded snpf() in process_socket()
  to match the rest of the file and prevent buffer overflow

+++ src/proc.c
@@ -244,7 +244,7 @@ void alloc_lfile(char *name, int num) {
  */
 
 void alloc_lproc(int pid, int pgid, int ppid, UID_ARG uid, char *cmd, int pss, int sel_flags) {
-    static int size = 0;
+    static size_t size = 0;
 
     if (!LocalProcTable) {
         if (!(LocalProcTable =
@@ -260,7 +260,7 @@ void alloc_lproc(int pid, int pgid, int ppid, UID_ARG uid, char *cmd, int pss, i
         tmp = (struct lproc *)realloc((MALLOC_P *)LocalProcTable,
                                       (MALLOC_S)(size * sizeof(struct lproc)));
         if (!tmp) {
-            fprintf(stderr, "%s: no realloc space for %d local proc structures\n", ProgramName,
+            fprintf(stderr, "%s: no realloc space for %zu local proc structures\n", ProgramName,
                     size);
             free((FREE_P *)LocalProcTable);
             LocalProcTable = NULL;
